feat - Creating Security module for API
This module includes an encryption and decryption strategy to handles with cryptograph and decryptograph fields that is sent or received in the API.
This ensure that sensitive data will not be exposed in unsecure connection.
Finally, this strategies utilize ECDH keys, that is at the same time more light to be generated and transited and also offers the same security level as the 7680-bit RSA key, since I'm using the P384 ECDH key.